+++ react/components/QuoteDetails.tsx
@@ -5,6 +5,7 @@ import { useIntl, defineMessages, FormattedMessage } from 'react-intl'
 import { useQuery, useMutation } from 'react-apollo'
 import { useRuntime } from 'vtex.render-runtime'
 import {
+  Alert,
   Layout,
   PageHeader,
   PageBlock,
@@ -106,6 +107,21 @@ const messages = defineMessages({
   addNote: {
     id: `${storePrefix}quote-details.add-note.label`,
   },
+  percentageDiscount: {
+    id: `${storePrefix}quote-details.percentage-discount.title`,
+  },
+  quotedSubtotal: {
+    id: `${storePrefix}quote-details.quoted-subtotal.title`,
+  },
+  originalSubtotal: {
+    id: `${storePrefix}quote-details.original-subtotal.title`,
+  },
+  quotePrice: {
+    id: `${storePrefix}quote-details.quote-price.title`,
+  },
+  originalPrice: {
+    id: `${storePrefix}quote-details.original-price.title`,
+  },
 })
 
 const QuoteDetails: FunctionComponent = () => {
@@ -181,6 +197,8 @@ const QuoteDetails: FunctionComponent = () => {
   const [discountState, setDiscountState] = useState(0)
   const [updatingQuoteState, setUpdatingQuoteState] = useState(false)
   const [usingQuoteState, setUsingQuoteState] = useState(false)
+  const [originalSubtotal, setOriginalSubtotal] = useState(0)
+  const [updatingSubtotal, setUpdatingSubtotal] = useState(0)
 
   const { data, loading, refetch } = useQuery(GET_QUOTE, {
     variables: { id: params?.id },
@@ -195,6 +213,24 @@ const QuoteDetails: FunctionComponent = () => {
     data: permissionsData,
     loading: permissionsLoading,
   } = useQuery(GET_PERMISSIONS, { ssr: false })
+
+  useEffect(() => {
+    if (!quoteState.items.find((item) => item.error)) {
+      setUpdatingSubtotal(
+        quoteState.items.reduce(
+          (sum, item) => sum + item.sellingPrice * item.quantity,
+          0
+        )
+      )
+    }
+
+    const price = quoteState.items.reduce(
+      (sum: number, item: QuoteItem) => sum + item.listPrice * item.quantity,
+      0
+    )
+
+    setOriginalSubtotal(price)
+  }, [quoteState])
 
   useEffect(() => {
     if (!orderFormData?.orderForm) return
@@ -242,6 +278,7 @@ const QuoteDetails: FunctionComponent = () => {
       },
     } = data
 
+    setUpdatingSubtotal(subtotal)
     setQuoteState({
       id,
       costCenter,
@@ -277,7 +314,7 @@ const QuoteDetails: FunctionComponent = () => {
 
   const handleSaveQuote = () => {
     setUpdatingQuoteState(true)
-    const { id, items, subtotal } = quoteState
+    const { id, items } = quoteState
 
     const itemsChanged = !arrayShallowEqual(data.getQuote.items, items)
 
@@ -285,7 +322,7 @@ const QuoteDetails: FunctionComponent = () => {
       variables: {
         id,
         ...(itemsChanged && { items }),
-        ...(itemsChanged && { subtotal }),
+        ...(itemsChanged && { subtotal: updatingSubtotal }),
         ...(noteState && { note: noteState }),
         decline: false,
       },
@@ -354,18 +391,17 @@ const QuoteDetails: FunctionComponent = () => {
   const handleUpdateSellingPrice: (
     id: string
   ) => ChangeEventHandler<HTMLInputElement> = (itemId) => (event) => {
-    let newSubtotal = 0
     const newItems = quoteState.items.map((item: QuoteItem) => {
       if (item.id === itemId) {
-        newSubtotal += +event.target.value * item.quantity * 100
+        const newPrice = ((event.target.value as unknown) as number) * 100
 
         return {
           ...item,
-          sellingPrice: ((event.target.value as unknown) as number) * 100,
+          sellingPrice: newPrice,
+          error:
+            !newPrice || newPrice / item.listPrice < maxDiscountState / 100,
         }
       }
-
-      newSubtotal += item.sellingPrice * item.quantity
 
       return item
     })
@@ -373,7 +409,6 @@ const QuoteDetails: FunctionComponent = () => {
     setQuoteState({
       ...quoteState,
       items: newItems,
-      subtotal: newSubtotal,
     })
   }
 
@@ -396,6 +431,7 @@ const QuoteDetails: FunctionComponent = () => {
       return item
     })
 
+    setUpdatingSubtotal(newSubtotal)
     setQuoteState({
       ...quoteState,
       items: newItems,
@@ -410,7 +446,7 @@ const QuoteDetails: FunctionComponent = () => {
 
     items.forEach((item: QuoteItem) => {
       const newSellingPrice = Math.round(
-        item.sellingPrice * ((100 - percent) / 100)
+        item.listPrice * ((100 - percent) / 100)
       )
 
       newSubtotal += newSellingPrice * item.quantity
@@ -418,6 +454,7 @@ const QuoteDetails: FunctionComponent = () => {
       newItems.push({ ...item, sellingPrice: newSellingPrice })
     })
 
+    setUpdatingSubtotal(newSubtotal)
     setQuoteState({
       ...quoteState,
       items: newItems,
@@ -506,8 +543,18 @@ const QuoteDetails: FunctionComponent = () => {
                   <Table
                     totalizers={[
                       {
-                        label: formatMessage(messages.subtotal),
-                        value: formatPrice(quoteState.subtotal),
+                        label: formatMessage(messages.originalSubtotal),
+                        value: formatPrice(originalSubtotal),
+                      },
+                      {
+                        label: formatMessage(messages.percentageDiscount),
+                        value: `${Math.round(
+                          100 - (quoteState.subtotal / originalSubtotal) * 100
+                        )}%`,
+                      },
+                      {
+                        label: formatMessage(messages.quotedSubtotal),
+                        value: formatPrice(updatingSubtotal),
                       },
                       {
                         label: formatMessage(messages.expiration),
@@ -526,7 +573,6 @@ const QuoteDetails: FunctionComponent = () => {
                         ),
                       },
                     ]}
-                    disableHeader
                     fullWidth
                     schema={{
                       properties: {
@@ -573,13 +619,35 @@ const QuoteDetails: FunctionComponent = () => {
                           },
                           minWidth: 300,
                         },
+                        listPrice: {
+                          title: formatMessage(messages.originalPrice),
+                          headerRight: true,
+                          width: 120,
+                          cellRenderer: ({ rowData }: any) => {
+                            return (
+                              isSalesRep && (
+                                <div
+                                  className={`w-100 tr${
+                                    rowData.listPrice !== rowData.sellingPrice
+                                      ? ' strike '
+                                      : ''
+                                  }`}
+                                >
+                                  <FormattedCurrency
+                                    value={rowData.listPrice / 100}
+                                  />
+                                </div>
+                              )
+                            )
+                          },
+                        },
                         sellingPrice: {
-                          title: formatMessage(messages.price),
+                          title: formatMessage(messages.quotePrice),
                           headerRight: true,
                           width: 120,
                           cellRenderer: ({
                             cellData: sellingPrice,
-                            rowData: { id: itemId },
+                            rowData: { id: itemId, error },
                           }: any) => {
                             if (
                               formState.isEditable &&
@@ -593,6 +661,7 @@ const QuoteDetails: FunctionComponent = () => {
                                   onChange={handleUpdateSellingPrice(itemId)}
                                   currencyCode={currencyCode}
                                   locale={locale}
+                                  error={error}
                                 />
                               )
                             }
@@ -652,6 +721,7 @@ const QuoteDetails: FunctionComponent = () => {
                       <h3 className="t-heading-4 mb8">
                         <FormattedMessage id="store/b2b-quotes.quote-details.apply-discount.title" />
                       </h3>
+
                       <Slider
                         onChange={([value]: [number]) => {
                           handlePercentageDiscount(value)
@@ -665,6 +735,7 @@ const QuoteDetails: FunctionComponent = () => {
                         formatValue={(a: number) => `${a}%`}
                         value={discountState}
                       />
+
                       <div className="mt1">
                         <FormattedMessage id="store/b2b-quotes.quote-details.apply-discount.help-text" />
                       </div>
@@ -687,7 +758,7 @@ const QuoteDetails: FunctionComponent = () => {
                       return (
                         <div key={index} className="ph4 pv2">
                           <Card>
-                            <p>
+                            <div>
                               <FormattedMessage
                                 id="store/b2b-quotes.quote-details.update-history.update-details"
                                 values={{
@@ -708,15 +779,15 @@ const QuoteDetails: FunctionComponent = () => {
                                   index,
                                 }}
                               />
-                            </p>
+                            </div>
                             {update.note && (
-                              <p>
+                              <div>
                                 <b>
                                   <FormattedMessage id="store/b2b-quotes.quote-details.update-history.notes" />
                                 </b>
                                 <br />
                                 {update.note}
-                              </p>
+                              </div>
                             )}
                           </Card>
                         </div>
@@ -743,14 +814,25 @@ const QuoteDetails: FunctionComponent = () => {
                     </div>
                   )}
                 </div>
+
+                {quoteState.items.some((item) => item.error) ? (
+                  <div className="mb4">
+                    <Alert type="error">
+                      <FormattedMessage
+                        id="store/b2b-quotes.quote-details.discount-error"
+                        values={{ count: noteState.length }}
+                      />
+                    </Alert>
+                  </div>
+                ) : null}
+
                 <div className="nowrap">
                   {quoteDeclinable && (
                     <span className="mr4">
                       <Button
                         variation="danger"
                         onClick={() => handleDeclineQuote()}
-                        isLoading={updatingQuoteState}
-                        disabled={!formState.isEditable}
+                        disabled={!formState.isEditable || updatingQuoteState}
                       >
                         <FormattedMessage id="store/b2b-quotes.quote-details.decline" />
                       </Button>
@@ -762,9 +844,10 @@ const QuoteDetails: FunctionComponent = () => {
                       onClick={() => handleSaveQuote()}
                       isLoading={updatingQuoteState}
                       disabled={
-                        quoteState.items.length &&
-                        noteState === '' &&
-                        arrayShallowEqual(items, quoteState.items)
+                        quoteState.items.some((item) => item.error) ||
+                        (quoteState.items.length &&
+                          noteState === '' &&
+                          arrayShallowEqual(items, quoteState.items))
                       }
                     >
                       {isSalesRep ? (